nemo_punctuation_capitalization/punctcap_service/requirements.txt
misc-utils@git+https://github.com/dertilo/misc-utils.git#egg=misc-utils
python-multipart
nemo_toolkit[nlp]
uvicorn[standard]
fastapi